+++ library/src/main/java/cloud/pace/sdk/utils/DeviceUtils.kt
@@ -112,8 +112,13 @@ object DeviceUtils {
         return string
     }
 
-    fun getRedirectScheme(context: Context): String? {
+    fun getPACERedirectScheme(context: Context): String? {
         val applicationInfo = context.packageManager?.getApplicationInfo(context.packageName, PackageManager.GET_META_DATA)
         return applicationInfo?.metaData?.get("pace_redirect_scheme")?.toString()
+    }
+
+    fun getAppAuthRedirectScheme(context: Context): String? {
+        val applicationInfo = context.packageManager?.getApplicationInfo(context.packageName, PackageManager.GET_META_DATA)
+        return applicationInfo?.metaData?.get("appAuthRedirectScheme")?.toString()
     }
 }